I changed the thread organization, but don't see a difference in the behavior.
As it is, you still have to CTRL+C after quitting the GUI to get the data taking loop to actually stop.
I read online that a thread will keep running until it stops itself, so I would have to figure out how to get the data taking thread to kill itself when the GUI stops.